+++ tests/src/testcases.js
@@ -134,6 +134,15 @@ module.exports = testCases = (client, db = false) => {
       }
     },
 
+    // Send a custom stanza message to room or a direct chat
+    customStanza: (to, callback) => {
+      self.config.customStanza = true;
+      self.message(to, () => {
+        callback();
+        self.config.customStanza = false;
+      });
+    },
+
     // Send message to room
     mucMessage: (room, callback) => {
       // Make sure we join to the given room first, before sending the message
@@ -145,9 +154,24 @@ module.exports = testCases = (client, db = false) => {
         `Send a new message from ${self.coloredJid(self.config.jid.local)} ` +
         `to ${self.coloredJid(room)}`
       );
-      let message = faker.hacker.phrase();
       let callbacked = false;
+      let message = faker.hacker.phrase();
       let safeMessage = self.utils.escapeXml(message);
+      let opts = { body: message };
+
+      if (self.config.customStanza) {
+        delete opts.body;
+        opts.store = true;
+        opts.customStanza = {
+          id: 4,
+          message: "Max's new car is great!",
+          owner: {
+            id: 7,
+            name: 'Max Mustermann'
+          }
+        };
+        safeMessage = '<very-custom-stanza ';
+      }
 
       self.config.lastMessage = safeMessage;
       self.utils.setMatcher(msgMatcher(safeMessage), (xml, direction) => {
@@ -155,17 +179,20 @@ module.exports = testCases = (client, db = false) => {
         callbacked = true;
 
         // Check the input message
-        stanza.message(roomJid, self.config.jid.local)(xml, direction);
+        stanza.message(
+          roomJid,
+          self.config.jid.local,
+          !('customStanza' in opts)
+        )(xml, direction);
         // Signalize that the message sending is done
         callback && callback();
       });
 
       // Send the actual message
-      self.client.sendMessage({
+      self.client.sendMessage(Object.assign({
         to: roomJid,
-        body: message,
         type: 'groupchat'
-      });
+      }, opts));
     },
 
     // Send direct message to another user
@@ -176,19 +203,31 @@ module.exports = testCases = (client, db = false) => {
         `Send a new message from ${self.coloredJid(self.config.jid.local)} ` +
         `to ${to.blue}`
       );
-      let message = faker.hacker.phrase();
       let callbacked = false;
+      let message = faker.hacker.phrase();
       let safeMessage = self.utils.escapeXml(message);
+      let opts = { body: message };
+
+      if (self.config.customStanza) {
+        delete opts.body;
+        opts.store = true;
+        opts.customStanza = {
+          id: 4,
+          message: "Max's new car is great!",
+          owner: {
+            id: 7,
+            name: 'Max Mustermann'
+          }
+        };
+        safeMessage = '<very-custom-stanza ';
+      }
 
       // Unfortunately XMPP does not respond anything for direct messages, so
       // we must check the database - but first wait a bit to allow ejabberd to
       // write the message to the database
       database.archiveOneMoreDirect(callback, (callback) => {
         // Send the actual message
-        self.client.sendMessage({
-          to: toJid,
-          body: message
-        });
+        self.client.sendMessage(Object.assign({ to: toJid }, opts));
         setTimeout(callback, messageTimeout);
       });
     },
@@ -550,7 +589,8 @@ module.exports = testCases = (client, db = false) => {
             body: res.body,
             conversation: res.from.bare,
             from: res.from.resource,
-            unread: res.unread == true
+            unread: res.unread == true,
+            customStanza: res.customStanza
           };
           return memo;
         }, {});
@@ -602,6 +642,79 @@ module.exports = testCases = (client, db = false) => {
                 self.utils.log(what, false, self.config.testLevel + 2);
               }
             });
+
+            switchDone(callback);
+          });
+        });
+      };
+    },
+
+    // Check the message archive for the given conversation and user - we use
+    // the first message of the archive to validate the full message payload
+    // (custom stanza)
+    expectMamHistoryCustomStanza: (user, conversation) => {
+      return (callback) => {
+        self.switchUser(user, (test, switchDone) => {
+          test.queryMamHistory(conversation, (messages) => {
+            let message = messages[Object.keys(messages)[0]];
+            let errors = 0;
+
+            if (!message || !message.customStanza) {
+              self.utils.logError(
+                `No message found or no custom stanza included ` +
+                `at ${conversation} ` +
+                `of ${user}`,
+                ''
+              );
+              return switchDone(callback);
+            }
+
+            if (message.customStanza.id != '4') {
+              errors++;
+              self.utils.logError(
+                `Unexpected custom stanza (id)`,
+                '',
+                `${message.customStanza.id} (actual) != ` +
+                `4 (expected)`
+              );
+            }
+
+            if (message.customStanza.message != "Max's new car is great!") {
+              errors++;
+              self.utils.logError(
+                `Unexpected custom stanza (message)`,
+                '',
+                `${message.customStanza.message} (actual) != ` +
+                `"Max's new car is great!" (expected)`
+              );
+            }
+
+            if (message.customStanza.owner.id != '7') {
+              errors++;
+              self.utils.logError(
+                `Unexpected custom stanza (owner.id)`,
+                '',
+                `${message.customStanza.owner.id} (actual) != ` +
+                `7 (expected)`
+              );
+            }
+
+            if (message.customStanza.owner.name != 'Max Mustermann') {
+              errors++;
+              self.utils.logError(
+                `Unexpected custom stanza (owner.name)`,
+                '',
+                `${message.customStanza.owner.name} (actual) != ` +
+                `"Max Mustermann" (expected)`
+              );
+            }
+
+            if (errors == 0) {
+              self.utils.log(
+                `Check custom stanza completeness on message`,
+                false, self.config.testLevel + 2
+              );
+            }
 
             switchDone(callback);
           });